refactor: 로그인, 회원가입 관련 User->Auth 패키지로 이동
